refactor: modules, components, assets, migrate to arrow fn
