--- src/Hero/Hero.css
@@ -1,10 +1,0 @@
-.hero {
-  width: 100%;
-  height: 90vh;
-  background: linear-gradient(0deg, rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4)),
-    url(hero-background.jpeg);
-  background-blend-mode: darken, normal;
-
-  color: white;
-  font-size: x-large;
-}
