+++ src/Hero/Hero.tsx
@@ -1,7 +1,18 @@
-import "./Hero.css";
+import styled from "styled-components";
+import heroBackground from "src/assets/images/backgrounds/hero-background.jpeg";
 
-function Hero() {
-  return <div className="hero">Hero v2</div>;
-}
+const HeroContainer = styled.div`
+  width: 100%;
+  height: 90vh;
+  color: white;
+  font-size: x-large;
+  background: linear-gradient(0deg, rgba(0, 0, 0, 0.4), rgba(0, 0, 0, 0.4)),
+    url(${heroBackground});
+  background-blend-mode: darken, normal;
+`;
+
+const Hero: React.FC = () => {
+  return <HeroContainer>Hero v2</HeroContainer>;
+};
 
 export default Hero;
